+++ test/td_bg/business_concepts/audit_test.exs
@@ -17,7 +17,10 @@ defmodule TdBg.BusinessConcepts.AuditTest do
   end
 
   setup do
-    stub(MockClusterHandler, :call, fn :ai, TdAi.Indices, :exists_enabled?, [] ->
+    stub(MockClusterHandler, :call, fn :ai,
+                                       TdAi.Indices,
+                                       :exists_enabled?,
+                                       [[index_type: "suggestions"]] ->
       {:ok, true}
     end)