+++ test/td_bg/business_concepts/business_concept_bulk_update_test.exs
@@ -78,7 +78,10 @@ defmodule TdBg.BusinessConceptBulkUpdateTest do
       id: "999"
     })
 
-    stub(MockClusterHandler, :call, fn :ai, TdAi.Indices, :exists_enabled?, [] ->
+    stub(MockClusterHandler, :call, fn :ai,
+                                       TdAi.Indices,
+                                       :exists_enabled?,
+                                       [[index_type: "suggestions"]] ->
       {:ok, true}
     end)