+++ test/td_bg_web/controllers/search_controller_test.exs
@@ -4,6 +4,8 @@ defmodule TdBgWeb.SearchControllerTest do
   alias TdCluster.TestHelpers.TdAiMock
   alias TdCore.Search.IndexWorkerMock
 
+  @index_type "suggestions"
+
   describe "/api/business_concepts/search/embeddings/_put" do
     setup do
       IndexWorkerMock.clear()
@@ -13,7 +15,7 @@ defmodule TdBgWeb.SearchControllerTest do
 
     @tag authentication: [role: "admin"]
     test "triggers a put embeddings action", %{conn: conn} do
-      TdAiMock.Indices.exists_enabled?(&Mox.expect/4, {:ok, true})
+      TdAiMock.Indices.exists_enabled?(&Mox.expect/4, [index_type: @index_type], {:ok, true})
 
       assert conn
              |> post(Routes.search_path(conn, :embeddings, %{}))
@@ -24,7 +26,7 @@ defmodule TdBgWeb.SearchControllerTest do
 
     @tag authentication: [role: "admin"]
     test "returns forbiddend when there are no indices enabled", %{conn: conn} do
-      TdAiMock.Indices.exists_enabled?(&Mox.expect/4, {:ok, false})
+      TdAiMock.Indices.exists_enabled?(&Mox.expect/4, [index_type: @index_type], {:ok, false})
 
       assert conn
              |> post(Routes.search_path(conn, :embeddings, %{}))