+++ test/td_bg_web/controllers/shared_domains_controller_test.exs
@@ -10,7 +10,10 @@ defmodule TdBgWeb.SharedDomainControllerTest do
   end
 
   setup do
-    stub(MockClusterHandler, :call, fn :ai, TdAi.Indices, :exists_enabled?, [] ->
+    stub(MockClusterHandler, :call, fn :ai,
+                                       TdAi.Indices,
+                                       :exists_enabled?,
+                                       [[index_type: "suggestions"]] ->
       {:ok, true}
     end)